requirements_dev.txt
prefect
pydantic
pymysql
sqlalchemy
pyyaml
dash
pandas
plotly
numpy
coolname
datapane
pytest
coverage
pytest-mock-resources[mysql]
freezegun